requirements.txt
Flask==2.3.3
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
PyPDF2==3.0.1
python-docx==0.8.11
Werkzeug==2.3.7
gunicorn==21.2.0
sentence-transformers==3.3.1